Vehicle Intercom first draft
C++17 Support
Resolving some TODO's
Typo Fixes
Fixed radioReplaceProcess not spawned
Fixed Radios not being replaced.. durp
Fixed Typo
Fixed PFH's running after mission end (Arma bug)
Moved MISSIONEND EH to be handles in pipe extension instead of TS Plugin
Moved isOverRadio functions to clientData
Fixed a memory leak in processStereoEffect
Improved parsing of Game Commands to increase performance in dertermining which command was sent (over 200% more to come)
Added debug Executable for easy performance testing
Added new stringSplit function using boost::string_ref
Added OutputDebugString Logger class
Added C++17 support using VisualStudio 15 features
Rewrote versionNumber comparison
Added Cache for Teamspeak Channel Names
Merged feature/pluginOptimizations into 1.0

+++ ts/src/Clunk.h
@@ -8,7 +8,7 @@ class Clunk
 public:
 
 
-	void process(short * samples, int channels, int sampleCount, Direction3D pos, AngleRadians direction)
+	void process(short * samples, uint8_t channels, size_t sampleCount, Direction3D pos, AngleRadians direction)
 	{
 		float x, y, z;
 		std::tie(x, y, z) = pos.get();
@@ -25,7 +25,7 @@ public:
 
 		input_buffer.insert(input_buffer.end(), samples, samples + (sampleCount*channels));
 
-		if (input_buffer.size() > clunk::Hrtf::WINDOW_SIZE * channels * 2)
+		if (input_buffer.size() > clunk::Hrtf::WINDOW_SIZE * channels * 2u)
 		{
 
 			const int to_process = (int) input_buffer.size();
